+++ bin/header.hpp
@@ -1,0 +1,32 @@
+/*
+# ===============================================================
+# About the code
+# ===============================================================
+# This section contains all the necessary libraries used in this
+# software package. 
+#
+# This method is implemented to prevent unecessary library 
+# declarations that have already been delcared.
+# 
+# ===============================================================
+*/
+
+#ifndef header_hpp
+#define header_hpp
+
+// Standard libraries
+#include <string>
+#include <iostream>
+#include <vector>
+#include <cmath>
+#include <assert.h>
+#include <regex>
+#include <algorithm>
+#include <chrono>
+
+// Custom libaries
+#include "str_handler.hpp"
+#include "Eigen/Core"
+#include "Eigen/Dense" // We use dense matrix initially
+
+#endif