+++ bin/main.sh
@@ -1,0 +1,1 @@
+g++ circuit.cpp ../bin/components/branch.cpp ../bin/components/resistor.cpp ../bin/components/vsource.cpp